Add the DATETIME_API macro on the operator << of the class datetime and solved a potential bug on a conversion from size_t to int.

+++ src/datetime.cpp
@@ -462,7 +462,7 @@ namespace jed_utils
 		return datetime(year, month, day, hour, minute, second, day_period);
 	}
 
-	int datetime::_parse_intvalue(const string &pattern, int index, int mask_length, const string &parse_str)
+	int datetime::_parse_intvalue(const string &pattern, int index, size_t mask_length, const string &parse_str)
 	{
 		long converted_value;
 		int ret_val;
